requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0